+++ aldrin/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `UnknownCall`, which represents an unknown pending call.
 
+### Changed
+
+- Remove `RequiredFieldMissing` error.
+
 ## [0.10.0] - 2024-11-26
 
 ### Added